requirements.txt
toml~=0.10.2
python-dotenv~=1.0.0
chromadb~=0.3.27
tiktoken~=0.4.0
pypdf~=3.12.1
transformers~=4.30.2
council-ai==0.0.3
sentence-transformers==2.2.2